Update arangosearch-views-reference.md
Fix "higher" and "lower" in explanation for commitIntervalMsec

A higher value for the interval means the data is commited less frequently, so there will be more time where the index does not account for the changes. Also, memory consumption will grow because the data is not yet flushed to disk yet.