DOC-609 | Fix aliasing and deprecation admonition
Only substitute version number with alias in internal links if currently using an alias in the URL (or if the version doesn't have a proper alias, i.e. name == alias).

Perform the replacement for all versions accordingly so that rare uses of cross-version links (like in the deprecation admonition) are substituted correctly. Previously, all links were changed to the current URL version.