Split code for copy to clipboard and adding the Show more button
This allows to independently call the latter, which we need to do for tabbed content for which we don't know the height until it is activated. The the event handler at a higher level so that it can handle buttons added after page load

Also remove some dead code for copy to clipboard.