DOC-660 | Improve logic for "Show more" (#517)
Don't use conflicting detection of long code blocks (element height vs. number of lines). Ensure we always reveal ~5 lines or don't collapse content. Also don't collapse code blocks already hidden by default ("Show output").

* Fix regression with hash-only links that shouldn't be aliased

An anchor link does not contain a slash to split by. The second element is thus undefined and it got spliced in, then turned into an extra slash, breaking the fragment identifier.

* Split code for copy to clipboard and adding the Show more button

This allows to independently call the latter, which we need to do for tabbed content for which we don't know the height until it is activated. The the event handler at a higher level so that it can handle buttons added after page load

Also remove some dead code for copy to clipboard.

* Use .on("click", ...) over the deprecated .click(...)

* Remove rounding from tab underline

* Apply code block styles also independently of copy to clipboard class

* CSS: Normalize 0px to 0