OpenAPI: Fix and improve validation, resolve conflicts in endpoint descriptions (#835)
* Attempt to trigger OpenAPI validation at the end of the build

Move invocation to home.navigation.html and set high output format weight, add locks and debug printing for map in arangoproxy

There still appear to be openapi render hook calls after the validation, and Hugo reports the end of the build way before that

* OpenAPI Render Hook: Substitute version alias with version number to prevent duplicate invocations

This is only an issue for endpoint descriptions that contain internal links that get replaced with absolute URLs for api-docs.json / SwaggerUI. Hugo's caching of resources.GetRemote prevents duplicate calls arangoproxy's /openapi if the requests are identical

* Replace invocation of OpenAPI validation

It no longer gets triggered in live builds (hugo server) but only static builds, where we can reliably start the validation after the build

* Add debugging and validation for OpenAPI endpoints to find duplicate operationIds and conflicting endpoint descriptions

Note that identical endpoint descriptions for the same version are not considered an error (used for a few View endpoints on purpose). In fact, Hugo's caching of resources.GetRemote prevents arangoproxy from even getting called twice, which we take advantage of.

* Fix newly discovered conflicts between arangosearch and search-alias View endpoints and operationIds

* Revert addition of output format weight, use single curl HTTP method

arangoproxy's endpoints work with any method (HEAD, POST, ...) but using -I (HEAD) has the advantage that it only returns the HTTP headers

* Propagate validation errors to start_hugo.sh to let the pipeline fail if needed

Also add a temporary bug to an endpoint description for testing

* Propagate spec errors to CI

* Test duplicate operationId

* Remove a required property to cause swagger-cli to fail

* Revert removal of required property

* Reset error booleans